requirement.txt
python-dotenv
llama-index
openai
cryptography
uvicorn